+++ src/layer-manager.js
@@ -1,4 +1,6 @@
 import isEmpty from 'lodash/isEmpty';
+import Promise from 'utils/promise';
+
 import LayerModel from './layer-model';
 
 function checkPluginProperties(plugin) {
@@ -38,61 +40,6 @@ class LayerManager {
     checkPluginProperties(this.plugin);
   }
 
-  /**
-   * Render layers
-   */
-  renderLayers() {
-    if (this.layers.length > 0) {
-      this.layers.forEach((layerModel) => {
-        const { changedAttributes } = layerModel;
-        const { sqlParams, params, layerConfig } = changedAttributes;
-        const hasChanged = Object.keys(changedAttributes).length > 0;
-        const shouldUpdate = sqlParams || params || layerConfig;
-
-        if (layerModel.mapLayer && !shouldUpdate) {
-          // If layer exists and didn't change don't do anything
-          if (!hasChanged) {
-            return false;
-          }
-
-          // In case has changed, just update it
-          if (hasChanged) {
-            return this.updateLayer(layerModel);
-          }
-        }
-
-        if (layerModel.mapLayer && shouldUpdate) {
-          this.updateLayer(layerModel);
-        }
-
-
-        // adds a new promise to `this.promises` every time it gets called
-        this.requestLayer(layerModel);
-
-        // reset changedAttributes
-        return layerModel.set('changedAttributes', {});
-      });
-
-
-      if (Object.keys(this.promises).length === 0) {
-        return Promise.resolve(this.layers);
-      }
-
-
-      return Promise
-        .all(Object.values(this.promises))
-        .then(() => {
-          this.promises = {};
-          return this.layers;
-        })
-        .catch(() => {
-          this.promises = {};
-        });
-    }
-
-    // By default it will return empty layers
-    return Promise.resolve(this.layers);
-  }
 
   /**
    * Add layers
@@ -100,7 +47,7 @@ class LayerManager {
    * @param {Object} layerOptions
    */
   add(
-    layers,
+    layer,
     layerOptions = {
       opacity: 1,
       visibility: true,
@@ -108,26 +55,18 @@ class LayerManager {
       interactivity: null
     }
   ) {
-    if (typeof layers === 'undefined') {
-      console.error('layers is required');
+    if (typeof layer === 'undefined') {
+      console.error('layer is required');
       return this;
     }
 
-    if (!Array.isArray(layers)) {
-      console.error('layers should be an array');
-      return this;
-    }
+    const layerModel = new LayerModel({ ...layer, ...layerOptions });
+    const { id } = layerModel;
 
-    layers.forEach((layer) => {
-      const existingLayer = this.layers.find(l => l.id === layer.id);
-      const nextModel = { ...layer, ...layerOptions };
+    if (this.layers.find(l => l.id === id)) return this.layers;
+    this.layers.push(layerModel);
 
-      if (existingLayer) {
-        existingLayer.update(nextModel);
-      } else {
-        this.layers.push(new LayerModel(nextModel));
-      }
-    });
+    this.requestLayer(layerModel);
 
     return this.layers;
   }
@@ -136,17 +75,18 @@ class LayerManager {
    * Updating a specific layer
    * @param  {Object} layerModel
    */
-  updateLayer(layerModel) {
+  update(id, changedProps) {
+    const layerModel = this.getLayerModel(id);
+    if (!layerModel || !layerModel.mapLayer) return;
+
+    layerModel.update(changedProps);
+
     const {
       opacity,
       visibility,
       zIndex,
-      params,
-      sqlParams,
-      decodeParams,
-      layerConfig,
-      events
-    } = layerModel.changedAttributes;
+      decodeParams
+    } = changedProps;
 
     if (typeof opacity !== 'undefined') {
       this.plugin.setOpacity(layerModel, opacity);
@@ -157,108 +97,64 @@ class LayerManager {
     if (typeof zIndex !== 'undefined') {
       this.plugin.setZIndex(layerModel, zIndex);
     }
-    if (typeof events !== 'undefined') {
-      this.setEvents(layerModel);
-    }
 
-    if (!isEmpty(layerConfig)) this.plugin.setLayerConfig(layerModel);
-    if (!isEmpty(params)) this.plugin.setParams(layerModel);
-    if (!isEmpty(sqlParams)) this.plugin.setParams(layerModel);
     if (!isEmpty(decodeParams)) this.plugin.setDecodeParams(layerModel);
   }
 
   /**
    * Remove a layer giving a Layer ID
-   * @param {Array} layerIds
+   * @param {String} id
    */
-  remove(layerIds) {
+  remove(id) {
     const layers = this.layers.slice(0);
-    const ids = Array.isArray(layerIds) ? layerIds : [layerIds];
 
-    ids.forEach((id) => {
-      this.requestCancel(id);
-    });
+    this.requestCancel(id);
 
-    this.layers.forEach((layerModel, index) => {
-      if (ids) {
-        if (ids.includes(layerModel.id)) {
-          this.plugin.remove(layerModel);
-          layers.splice(index, 1);
-        }
-      } else {
-        this.plugin.remove(layerModel);
-      }
-    });
+    const layerModel = this.getLayerModel(id);
 
-    this.layers = ids ? layers : [];
+    if (layerModel) {
+      this.plugin.remove(layerModel);
+    }
+
+    this.layers = layers.filter(l => l.id !== id);
   }
 
   getLayers() {
     return this.layers;
   }
 
+  getLayerModel(id) {
+    return this.layers.find(l => l.id === id);
+  }
+
   /**
    * A namespace to set opacity on selected layer
-   * @param {Array} layerIds
+   * @param {String} id
    * @param {Number} opacity
    */
-  setOpacity(layerIds, opacity) {
-    const layerModels = this.layers.filter(l => layerIds.includes(l.id));
-
-    if (layerModels.length) {
-      layerModels.forEach((lm) => {
-        this.plugin.setOpacity(lm, opacity);
-      });
-    } else {
-      console.error("Can't find the layer");
-    }
+  setOpacity(id, opacity) {
+    const layerModel = this.getLayerModel(id);
+    this.plugin.setOpacity(layerModel, opacity);
   }
 
   /**
    * A namespace to hide or show a selected layer
-   * @param {Array} layerIds
+   * @param {String} id
    * @param {Boolean} visibility
    */
-  setVisibility(layerIds, visibility) {
-    const layerModels = this.layers.filter(l => layerIds.includes(l.id));
-
-    if (layerModels.length) {
-      layerModels.forEach((lm) => {
-        this.plugin.setVisibility(lm, visibility);
-      });
-    } else {
-      console.error("Can't find the layer");
-    }
+  setVisibility(id, visibility) {
+    const layerModel = this.getLayerModel(id);
+    this.plugin.setVisibility(layerModel, visibility);
   }
 
   /**
    * A namespace to set z-index on selected layer
-   * @param {Array} layerIds
+   * @param {String} id
    * @param {Number} zIndex
    */
-  setZIndex(layerIds, zIndex) {
-    const layerModels = this.layers.filter(l => layerIds.includes(l.id));
-
-    if (layerModels.length) {
-      layerModels.forEach((lm) => {
-        this.plugin.setZIndex(lm, zIndex);
-      });
-    } else {
-      console.error("Can't find the layer");
-    }
-  }
-
-  /**
-   * A namespace to set events on selected layer
-   * @param  {Object} layerModel
-   */
-  setEvents(layerModel) {
-    const { events } = layerModel;
-
-    if (events && this.plugin.setEvents) {
-      // Let's leave the managment of event to the plugin
-      this.plugin.setEvents(layerModel);
-    }
+  setZIndex(id, zIndex) {
+    const layerModel = this.getLayerModel(id);
+    this.plugin.setZIndex(layerModel, zIndex);
   }
 
   requestLayer(layerModel) {
@@ -276,14 +172,15 @@ class LayerManager {
     // every render method returns a promise that we store in the array
     // to control when all layers are fetched.
     this.promises[layerModel.id] = method.call(this, layerModel).then((layer) => {
-      layerModel.set('mapLayer', layer);
+      const { _canceled: canceled } = this.promises[layerModel.id];
+      if (!canceled) {
+        layerModel.set('mapLayer', layer);
 
-      this.plugin.add(layerModel, this.layers);
-      this.plugin.setZIndex(layerModel, layerModel.zIndex);
-      this.plugin.setOpacity(layerModel, layerModel.opacity);
-      this.plugin.setVisibility(layerModel, layerModel.visibility);
-
-      this.setEvents(layerModel);
+        this.plugin.add(layerModel, this.layers);
+        this.plugin.setZIndex(layerModel, layerModel.zIndex);
+        this.plugin.setOpacity(layerModel, layerModel.opacity);
+        this.plugin.setVisibility(layerModel, layerModel.visibility);
+      }
     });
 
     return this;
@@ -293,10 +190,9 @@ class LayerManager {
     // Cancel previous/existing request
     if (
       this.promises[layerModelId]
-      && this.promises[layerModelId].isPending
-      && this.promises[layerModelId].isPending()
+      && this.promises[layerModelId].cancel
     ) {
-      this.promises[layerModelId].cancel('Cancelling layer method request');
+      this.promises[layerModelId].cancel();
     }
   }
 }
